bert/vector_retreival_gptver/requirements.txt
streamlit>=1.32.0
mysql-connector-python>=8.0.33
sentence-transformers>=2.2.2
deep-translator>=1.11.4
numpy>=1.23.5
scikit-learn>=1.1.3
pandas>=1.5.3
openai>=1.30.1
arxiv>=1.4.6


